requirements.txt
Flask==3.0.0
Pillow==10.1.0
midiutil==1.2.1
pydub==0.25.1
numpy==1.26.2
Werkzeug==3.0.1
